+++ src/NitroHttp.Cli/Views/Interfaces/IResponseView.cs
@@ -13,5 +13,5 @@ public interface IResponseView
     /// <param name="responseStatus">The HTTP response status code.</param>
     /// <param name="responseCount">The number of returned items.</param>
     /// <param name="responseSize">The response size in bytes.</param>
-    void Display(string requestUrl, string response, int responseStatus, int responseCount, int responseSize);
+    void Display(string requestUrl, string response, int responseStatus, int responseCount, long responseSize);
 }
